Update drivetrain implementation
- Removed unused functions
- Simplified logic
- Added enums for DriveMode, GearShift, etc.
- Added logic to enable and disable odometry
- Optimized operations
- Migrate commands

+++ src/main/java/frc/robot/utils/Convert.java
@@ -5,7 +5,7 @@ import frc.robot.Constants.kDriveTrain;
 
 public final class Convert {
     
-    public static double EncoderUnitsToInches(float encoderUnits){
+    public static double EncoderUnitsToInches(double encoderUnits){
         if(SmartDashboard.getBoolean("LowGear", true)){
             return encoderUnits / kDriveTrain.encoderCPR * kDriveTrain.lowGearConversionFactor * kDriveTrain.wheelCircumferenceInches;
         }
